chore: split mechanics into submodules
